:art: Create createVectorStoreNode factory function and refactor all unified vector stores
