--- packages/@n8n/nodes-langchain/nodes/vector_store/VectorStoreInMemory/retrieve/description.ts
@@ -1,17 +1,0 @@
-import type { INodeProperties } from 'n8n-workflow';
-
-export const retrieveOperationDescription: INodeProperties[] = [
-	// {
-	// 	displayName: 'Options',
-	// 	name: 'options',
-	// 	type: 'collection',
-	// 	placeholder: 'Add Option',
-	// 	default: {},
-	// 	options: [metadataFilterField],
-	// 	displayOptions: {
-	// 		show: {
-	// 			operation: ['retrieve'],
-	// 		},
-	// 	},
-	// },
-];
