--- packages/@n8n/nodes-langchain/nodes/vector_store/VectorStoreSupabase/insert/description.ts
@@ -1,15 +1,0 @@
-import type { INodeProperties } from 'n8n-workflow';
-
-export const insertOperationDescription: INodeProperties[] = [
-	{
-		displayName: 'Specify the document to load in the document loader sub-node',
-		name: 'notice',
-		type: 'notice',
-		default: '',
-		displayOptions: {
-			show: {
-				operation: ['insert'],
-			},
-		},
-	},
-];
